3_crew/equity_analyst/pyproject.toml
[project]
name = "equity_analyst"
version = "0.1.0"
description = "equity_analyst using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.6.1",
    "python-dotenv>=1.0.0,<2.0.0"
]

[project.scripts]
equity_analyst = "equity_analyst.main:run"
run_crew = "equity_analyst.main:run"
train = "equity_analyst.main:train"
replay = "equity_analyst.main:replay"
test = "equity_analyst.main:test"
run_with_trigger = "equity_analyst.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
